# swagger.yaml
openapi: 3.0.3
info:
  title: Ibrat Lead Calling System API
  description: |
    Complete backend API for AI-powered lead calling system with admin and operator panels
    
    ## 🚀 Quick Start Guide
    
    ### **1. Authentication Flow**
    1. **Register User**: `POST /api/auth/register` (admin only)
       - Use test credentials: `john.doe@ibrat.com` / `ibrat123`
    2. **Login**: `POST /api/auth/login` with same credentials
       - Copy JWT token from response
    3. **Use APIs**: Add `Authorization: Bearer <token>` header
    
    ### **2. Test Credentials**
    - **Admin**: `john.doe@ibrat.com` / `ibrat123`
    - **Salesperson**: `jane.smith@ibrat.com` / `ibrat123`
    
    ### **3. API Access**
    - **Admin Panel**: Full system access
    - **Operator Panel**: Work interface only
    - **Real-time**: WebSocket connections
    
    ### **4. Swagger Features**
    - Click "Try it out" to test APIs
    - Use "Authorize" button for JWT tokens
    - View request/response schemas
  version: 1.0.0
  contact:
    name: Backend Team
    email: backend@ibrat.com
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.ibrat.com
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token from /api/auth/login
        
        **How to use:**
        1. Register: `POST /api/auth/register`
        2. Login: `POST /api/auth/login` 
        3. Copy token from response
        4. Click "Authorize" button above
        5. Enter: `Bearer <your_token>`
        
        **Example:**
        ```
        Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjUwN2YxZjc3YmNmODZjZDc5OTQzOTAxMSIsIm5hbWUiOiJKb2huIERvZSIsImVtYWlsIjoiam9obi5kb2VAaWJyYXQuY29tIiwicm9sZSI6ImFkbWluIiwiaWF0IjoxNzM1NzE5NjAwLCJleHAiOjE3MzU3MjU2MDB9.example
        ```
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        phone:
          type: string
          example: "+1234567890"
        role:
          type: string
          enum: [admin, salesperson]
          example: "salesperson"
        isActive:
          type: boolean
          example: true
        isAvailable:
          type: boolean
          example: true
        lastLogin:
          type: string
          format: date-time
        callStats:
          type: object
          properties:
            totalCalls:
              type: number
              example: 150
            successfulCalls:
              type: number
              example: 120
            totalDuration:
              type: number
              example: 7200
    Lead:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439012"
        phone:
          type: string
          example: "+1234567890"
        name:
          type: string
          example: "Lead Name"
        email:
          type: string
          format: email
          example: "lead@example.com"
        status:
          type: string
          enum: [pending, claimed, calling, answered, transferred, failed]
          example: "pending"
        priority:
          type: string
          enum: [urgent, high, medium, low]
          example: "high"
        callAttempts:
          type: number
          example: 0
        assignedTo:
          $ref: '#/components/schemas/User'
        tags:
          type: array
          items:
            type: string
          example: ["hot", "interested"]
...
    CallInitiateRequest:
      type: object
      required:
        - leadId
      properties:
        leadId:
          type: string
          example: "507f1f77bcf86cd799439012"
...
paths:
...
  /api/operator/dashboard:
    get:
      tags: [Operator Panel]
      summary: Get operator dashboard
      description: Retrieve personal performance overview (salesperson only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Operator dashboard retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  dashboard:
                    type: object
                    properties:
                      personalStats:
                        type: object
                        properties:
                          totalCalls:
                            type: integer
                          todayCalls:
                            type: integer
                          answeredCalls:
                            type: integer
                          completedCalls:
                            type: integer
                      assignedLeads:
                        type: array
                        items:
                          $ref: '#/components/schemas/Lead'
                      availableLeads:
                        type: array
                        items:
                          $ref: '#/components/schemas/Lead'
  /api/operator/leads/available:
    get:
      tags: [Operator Panel]
      summary: Get available leads
      description: Browse leads available for claiming (salesperson only)
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: priority
          in: query
          schema:
            type: string
            enum: [urgent, high, medium, low]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Available leads retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  leads:
                    type: array
                    items:
                      $ref: '#/components/schemas/Lead'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      pages:
                        type: integer
  /api/operator/leads/claim:
    post:
      tags: [Operator Panel]
      summary: Claim a lead
      description: Claim a lead for personal work (salesperson only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - leadId
              properties:
                leadId:
                  type: string
                  example: "507f1f77bcf86cd799439012"
      responses:
        '200':
          description: Lead claimed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Lead claimed successfully"
                  lead:
                    type: object
                    properties:
                      id:
                        type: string
                      phone:
                        type: string
                      name:
                        type: string
                      priority:
                        type: string
                      status:
                        type: string
  /api/operator/calls/start:
    post:
      tags: [Operator Panel]
      summary: Start a call
      description: Initiate a call to a claimed lead (salesperson only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CallInitiateRequest'
      responses:
        '200':
          description: Call started successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Call started successfully"
                  call:
                    type: object
                    properties:
                      id:
                        type: string
                      lead:
                        type: string
                      status:
                        type: string
                      twilioCallSid:
                        type: string
  /api/operator/profile/availability:
    put:
      tags: [Operator Panel]
      summary: Update availability
      description: Set online/offline status (salesperson only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - isAvailable
              properties:
                isAvailable:
                  type: boolean
                  example: false
      responses:
        '200':
          description: Availability updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Availability updated successfully"
                  isAvailable:
                    type: boolean
...